feat(#1): add museum screen
